+++ src/renderer/images.ts
@@ -1,29 +1,11 @@
-import { dungeonsByZoneId } from '../main/constants'
+import { dungeonsByZoneId, classicArenas, retailArenas, raidEncountersById, retailBattlegrounds, classicBattlegrounds, specializationById } from '../main/constants'
 
 interface imageObject {
     [id: number]: string;
 }
 
-// Load the arena images. This expects files with the name <id>.jpg 
-// to exist in the appropriate assets folder. 
-const retailArenaIDs = [
-    1672, // Blade's Edge
-    617,  // Dalaran
-    1505, // Nagrand 
-    572,  // Ruins of Lordareon
-    2167, // Robodome
-    1134, // Tigers Peak
-    980,  // Tol'viron
-    1504, // Black Rook
-    2373, // Empyrean Domain
-    1552, // Ashamane's Fall
-    1911, // Mugambala
-    1825, // Hook Point
-    2509, // Maldraxxus
-    2547  // Enigma Crucible
-]
-
-let retailArena: imageObject = {};
+const retailArenaIDs = Object.keys(retailArenas).map(v => parseInt(v, 10));
+const retailArena: imageObject = {};
 
 for (const id of retailArenaIDs) {
     try {
@@ -34,17 +16,8 @@ for (const id of retailArenaIDs) {
     };
 }
 
-// Load the classic arena images. This expects files with the name <id>.jpg 
-// to exist in the appropriate assets folder. 
-// @@ classic arena images? i.e. blade's edge...
-const classicArenaIDs = [
-    572,
-    559,
-    617,
-    562,
-]
-
-let classicArena: imageObject = {};
+const classicArenaIDs = Object.keys(classicArenas).map(v => parseInt(v, 10));
+const classicArena: imageObject = {};
 
 for (const id of classicArenaIDs) {
     try {
@@ -55,8 +28,7 @@ for (const id of classicArenaIDs) {
     };
 }
 
-const dungeonIDs = Object.keys(dungeonsByZoneId).map(v=> parseInt(v, 10))
-
+const dungeonIDs = Object.keys(dungeonsByZoneId).map(v => parseInt(v, 10));
 let dungeon: imageObject = {};
 
 for (const id of dungeonIDs) {
@@ -68,92 +40,8 @@ for (const id of dungeonIDs) {
     };
 }
 
-// Load the raid encounter images. This expects files with the 
-// name <id>.jpg to exist in the appropriate assets folder.
-const raidIDs = [
-    // Castle Nathria
-    // Shadowlands Tier 1
-    2398, // "Shriekwing",
-    2418, // "Huntsman Altimor",
-    2404, // "Sun King's Salvation",
-    2405, // "Xy'mox",
-    2383, // "Hungering Destroyer",
-    2406, // "Inerva",
-    2412, // "Council of Blood",
-    2399, // "Sludgefist",
-    2417, // "Stone Legion Generals",
-    2407, // "Sire Denathrius"
-
-    // Sanctum of Domination
-    // Shadowlands Tier 2
-    2523, // "The Tarragrue",
-    2433, // "Eye of the Jailer",
-    2429, // "The Nine",
-    2432, // "Remnant of Ner'zhul",
-    2434, // "Souldrender Dormazain",
-    2430, // "Painsmith Raznal",
-    2436, // "Guardian of the First Ones",
-    2431, // "Fatescribe Roh-Kalo",
-    2422, // "Kel'Thuzad",
-    2435, // "Sylvanas Windrunner",
-
-    // Sepulcher of the First Ones
-    // Shadowlands Tier 3
-    2512, // Vigilant Guardian
-    2537, // Jailer
-    2539, // Luhvium
-    2540, // Desaugne
-    2542, // Skolex
-    2543, // Dreadlords
-    2544, // Pantheon
-    2546, // Anduin
-    2549, // Rygelon
-    2553, // Artificer
-    2529, // Halondrus
-
-    // Vault of the Incarnates
-    // Dragonflight Tier 1
-    2587, // Eranog
-    2639, // Terros
-    2590, // The Primal Council
-    2592, // Sennarth
-    2635, // Dathea
-    2605, // Kurog
-    2614, // Broodkeeper Diurna
-    2607, // Raszageth
-
-    // Naxxramas
-    // WOTLK Classic Tier 1
-    1107, // Anub
-    1110, // Faerlina
-    1116, // Maexxna
-    1118, // Patchwerk
-    1117, // Noth
-    1112, // Heigan
-    1115, // Lotheb
-    1113, // Razuvious
-    1109, // Gothik
-    1121, // Horsemen
-    1119, // Sapphiron
-    1120, // Thaddius
-    1114, // Kel'Thuzad
-    1111, // Grobbulus
-    1108, // Gluth
-
-    // Eye of Eternity
-    // WOTLK Classic Tier 1
-    734, // Malygos
-
-    // Obsidian Sanctum
-    // WOTLK Classic Tier 1
-    742, // Sarthartion
-
-    // Vault of Archavon
-    // WOTLK Classic Tier 1
-    772, // Sarthartion
-]
-
-let raid: imageObject = {};
+const raidIDs = Object.keys(raidEncountersById).map(v => parseInt(v, 10));
+const raid: imageObject = {};
 
 for (const id of raidIDs) {
     try {
@@ -164,90 +52,32 @@ for (const id of raidIDs) {
     };
 }
 
-// Load the battleground images. This expects files with the 
-// name <id>.jpg to exist in the appropriate assets folder.
-
-// Some of these have two entries either because there is 
-// two versions of the map (EOTS casual vs rated) or for no 
-// reason other than I don't have a good exhaustive list. 
-const retailBattlegroundIDs = [
-    2107, // Arathi Basin
-    1681, // Arathi Basin
-    30,   // Alterac Valley
-    761,  // Battle for Gilneas
-    1105, // Deepwind Gorge
-    2245, // Deepwind Gorge
-    566,  // EOTS
-    968,  // EOTS
-    628,  // IOC
-    727,  // Silvershard Mines
-    1803, // Seething Shore
-    998,  // Temple of Kotmogu
-    726,  // Twin Peaks
-    489,  // Warsong Gulch
-    2106, // Warsong Gulch
-]
-
-let retailBattlegrounds: imageObject = {};
+const retailBattlegroundIDs = Object.keys(retailBattlegrounds).map(v => parseInt(v, 10));
+const retailBattlegroundImages: imageObject = {};
 
 for (const id of retailBattlegroundIDs) {
     try {
-        retailBattlegrounds[id] = require(`../../assets/battlegrounds/${id}.jpg`);
+        retailBattlegroundImages[id] = require(`../../assets/battlegrounds/${id}.jpg`);
     }
     catch (e) {
         console.error(`[Images] Unable to load image resource that was expected to exist.\n`, e)
     };
 }
 
-// Load the battleground images. This expects files with the 
-// name <id>.jpg to exist in the appropriate assets folder.
-
-// Some of these have two entries either because there is 
-// two versions of the map (EOTS casual vs rated) or for no 
-// reason other than I don't have a good exhaustive list. 
-const classicBattlegroundIDs = [
-    30,	// "Alterac Valley",
-    529, //"Arathi Basin",
-    566, //"Eye of the Storm",
-    607, //"Strand of the Ancients",
-    489, //"Warsong Gulch",
-]
-
-let classicBattlegrounds: imageObject = {};
+const classicBattlegroundIDs = Object.keys(classicBattlegrounds).map(v => parseInt(v, 10));
+const classicBattlegroundImages: imageObject = {};
 
 for (const id of classicBattlegroundIDs) {
     try {
-        classicBattlegrounds[id] = require(`../../assets/battlegrounds/${id}.jpg`);
+        classicBattlegroundImages[id] = require(`../../assets/battlegrounds/${id}.jpg`);
     }
     catch (e) {
         console.error(`[Images] Unable to load image resource that was expected to exist.\n`, e)
     };
 }
 
-// Load the battleground images. This expects files with the name <id>.jpg 
-// to exist in the appropriate assets folder. For reference see: 
-// https://wowpedia.fandom.com/wiki/SpecializationID
-const specIDs = [
-    250, 251, 252,      // Death Knight
-    577, 581,           // Demon Hunter
-    102, 103, 104, 105, // Druid
-    253, 254, 255,      // Hunter
-    62, 63, 64,         // Mage
-    268, 270, 269,      // Monk
-    65, 66, 70,         // Paladin
-    256, 257, 258,      // Priest
-    259, 260, 261,      // Rogue
-    262, 263, 264,      // Shaman
-    265, 266, 267,      // Warlock
-    71, 72, 73,         // Warrior
-    1467, 1468,         // Evoker
-    0,                  // Fallback
-]
-
-// Set a 'not found' icon for defaulting to if everything goes wrong.
-// This is also what we default to over upgrade of the application when 
-// previous videos don't have the metadata to show the spec. 
-let spec: imageObject = { 
+const specIDs = Object.keys(specializationById).map(v => parseInt(v, 10));
+const spec: imageObject = { 
     0: require("../../assets/icon/wowNotFound.png") 
 };